auth

+++ article-delete.php
@@ -1,16 +1,22 @@
 <?php
 
-include "db.php";
+session_start();
 
-if(isset($_GET['id'])) {
-    $id = $_GET['id'];
-    $query = "DELETE FROM articles WHERE id='$id'";
-    $result = mysqli_query($conn, $query);
-    if($result) {
-        header("location: index.php");
+include 'db.php';
+
+if (isset($_SESSION['auth'])) {
+    if (isset($_GET['id'])) {
+        $id = $_GET['id'];
+        $query = "DELETE FROM articles WHERE id='$id'";
+        $result = mysqli_query($conn, $query);
+        if ($result) {
+            header('location: index.php');
+        } else {
+            echo mysqli_connect();
+        }
     } else {
-        echo mysqli_connect();
+        header('location: index.php');
     }
 } else {
-    header("location: index.php");
+    header('location: index.php');
 }
